// src/api/stetsom/model/pageBlock.ts
/**
 * Generated by orval v8.10.0 🍺
 * Do not edit manually.
 * Stetsom API
 * API documentation for Stetsom
 * OpenAPI spec version: 1.0.0
 */
import type { PageBlockData } from "./pageBlockData";
import type { PageBlockType } from "./pageBlockType";

export interface PageBlock {
  block_id: string;
  section_id: string;
  type: PageBlockType;
  /**
   * @minimum -9007199254740991
   * @maximum 9007199254740991
   */
  order: number;
  data: PageBlockData;
  available_locales: string[];
}